Fix UBSan and LSan issues
- Fix signed/unsigned conversion in FindVictim: change `total` from
  unsigned short to int since it intentionally goes negative
- Fix REMOVE_BIT macro to cast bit to typeof(var) before complement,
  avoiding signed int intermediate when used with unsigned long
- Fix FreeHates/FreeFears: were freeing wrong pointer (k->next instead of k)
- Add FreeHates/FreeFears calls to free_char to prevent memory leaks

+++ code/fight.c
@@ -1838,7 +1838,7 @@ struct char_data* FindVictim(struct char_data* ch) {
   unsigned short ptot = 0;
   unsigned short ktot = 0;
   unsigned short rtot = 0;
-  unsigned short total;
+  int total;
   unsigned short fjump = 0;
   unsigned short njump = 0;
   unsigned short cjump = 0;
@@ -2022,7 +2022,7 @@ struct char_data* FindAnyVictim(struct char_data* ch) {
   unsigned short ptot = 0;
   unsigned short ktot = 0;
   unsigned short rtot = 0;
-  unsigned short total;
+  int total;
   unsigned short fjump = 0;
   unsigned short njump = 0;
   unsigned short cjump = 0;
@@ -2411,7 +2411,7 @@ struct char_data* FindAnAttacker(struct char_data* ch) {
   unsigned short ctot = 0;
   unsigned short ntot = 0;
   unsigned short mtot = 0;
-  unsigned short total;
+  int total;
   unsigned short fjump = 0;
   unsigned short njump = 0;
   unsigned short cjump = 0;
